lightbox [nfc]: Edit code to prepare for lightbox refactor.
Changes:
- Renamed `src/lightbox/downloadImage.js` to `src/lightbox/download.js`.
- Changed default export to named export for `downloadImages`.
